Last but not least

+++ src/models/adminstration/box.screen.messages.model.ts
@@ -51,12 +51,12 @@ class BoxScreenMessagesModel {
   }
 
   // Get all BoxScreenMessages
-  async getAllBoxScreenMessages(): Promise<BoxScreenMessage[]> {
+  async getAllBoxScreenMessages(user: string): Promise<BoxScreenMessage[]> {
     const connection = await db.connect();
 
     try {
-      const sql = 'SELECT * FROM Box_Screen_Messages';
-      const result = await connection.query(sql);
+      const sql = 'SELECT * FROM Box_Screen_Messages WHERE user_id=$1';
+      const result = await connection.query(sql, [user]);
 
       return result.rows as BoxScreenMessage[];
     } catch (error) {
@@ -67,12 +67,16 @@ class BoxScreenMessagesModel {
   }
 
   // Get specific BoxScreenMessage by ID
-  async getBoxScreenMessageById(id: number): Promise<BoxScreenMessage> {
+  async getBoxScreenMessageById(
+    id: number,
+    user: string,
+  ): Promise<BoxScreenMessage> {
     const connection = await db.connect();
 
     try {
-      const sql = 'SELECT * FROM Box_Screen_Messages WHERE id=$1';
-      const result = await connection.query(sql, [id]);
+      const sql =
+        'SELECT * FROM Box_Screen_Messages WHERE id=$1 AND user_id=$2';
+      const result = await connection.query(sql, [id, user]);
 
       if (result.rows.length === 0) {
         throw new Error(`BoxScreenMessage with ID ${id} not found`);
@@ -94,8 +98,12 @@ class BoxScreenMessagesModel {
     const connection = await db.connect();
     try {
       // Check if the BoxScreenMessage exists
-      const checkSql = 'SELECT * FROM Box_Screen_Messages WHERE id=$1';
-      const checkResult = await connection.query(checkSql, [id]);
+      const checkSql =
+        'SELECT * FROM Box_Screen_Messages WHERE id=$1 AND user_id=$2';
+      const checkResult = await connection.query(checkSql, [
+        id,
+        boxScreenMessage.user_id,
+      ]);
 
       if (checkResult.rows.length === 0) {
         throw new Error(`BoxScreenMessage with ID ${id} does not exist`);
@@ -140,13 +148,16 @@ class BoxScreenMessagesModel {
   }
 
   // Delete BoxScreenMessage
-  async deleteBoxScreenMessage(id: number): Promise<BoxScreenMessage> {
+  async deleteBoxScreenMessage(
+    id: number,
+    user: string,
+  ): Promise<BoxScreenMessage> {
     const connection = await db.connect();
 
     try {
-      const sql = `DELETE FROM Box_Screen_Messages WHERE id=$1 RETURNING *`;
+      const sql = `DELETE FROM Box_Screen_Messages WHERE id=$1 AND user_id=$2 RETURNING *`;
 
-      const result = await connection.query(sql, [id]);
+      const result = await connection.query(sql, [id, user]);
 
       if (result.rows.length === 0) {
         throw new Error(`Could not find BoxScreenMessage with ID ${id}`);
